Add functionality to disable the power on LED.
This has only be tested for the Jumper T-Lite and will probably not work on other radios - especially with different screens.
Adaptations would need to be made in the menus for the other screens.
Also the blue LED in code is acutally the green on the Jumper T-Lite and the other way around.